MidiControlDialog: introduce Settings tab
this one will contain all MIDI configurations which can take effect without restarting the MIDI driver.

One of the biggest drawbacks of the MIDI tab of the `PreferencesDialog` was that for most settings it was not clear whether they do affect the MIDI input, output, or both. In this Settings tab we make two distinct columns to leave no room for interpretation