MidiBaseDriver: rework handling of message summaries
Previously, I used a `std::deque` containing raw `MidiInput::HandledInput` and `MidiOutput::HandledOutput` structs. But with the GUI thread reading and displaying those objects while the MIDI/AudioEngine threads rotates the deque, this was not designed in a parallel-prove way.

Instead, we now a mutex for each set of handlers and wrap the structs in `std::shared_ptr`.