Preferences: move temporary OSC port into OscServer
the OSC port provided via the CLI was introduced in `Preferences` as a convenience in order to not pass additional arguments to the `Hydrogen::create_instance()` method. But nowadays it turns out to be harmful. When we switch between different NSM sessions, a new set of preferences is read from the corresponding file and apply. This would make Hydrogen loose the previous port number. We now store it in `OscServer` instead.