PreferencesDialog: rework property reset on rejection
previously, the whole `Preferences` were reset by read the current user-level hydrogen.conf file and using it as a new instance. But this was a less than ideal approach. There are some settings, like the ones in the `MidiControlDialog`, window geometry, grid resolution or widget visibility, which are stored in `Preferences` but are not exposed in `PreferencesDialog`. Those were lost as well.

This patch resets only those setting group the user did modify. New settings have to be included in `PreferencesDialog::onRejected` as well. This is not convenient but we do not change/add settings on a monthly basis but very rarely. So, I think this way is fine