FakeDriver: rework to be a proper audio driver
previously, this class did not have a dedicated thread. Instead, it relied on being in song mode with looping disabled and some special handlers in `AudioEngine::processAudio` to quit the callback at the end of the song. Although, this was never used by the time I was maintainer.

Instead, `FakeDriver` does now have a dedicated worker thread calling the audio process callback and simulates way more properly an audio driver.