MidiActionManager: add worker thread
in an effort to handle MIDI events in a realtime-ish fashion, `MidiActionManager` has now a separate worker thread. This one is doing all the heavy lifting while incoming MIDI and OSC events only cause actions to be enqueued for processing by the worker thread.

The MIDI feedback and GUI handling of `MidiAction`, however, is still done synchronously via a different method