Time: add highResolutionSleep
since `std::this_thread::sleep_for` is not precise enough, we need some custom (platform-dependent) code to allow to high precision and real time handling.

`highResolutionSleep` calls the above routine of the C++ std library with a duration of a few seconds less than the amount it should actually sleep. The remaining milliseconds it checks time in a while loop till the entire duration has passed. The latter part is very inefficient and is meant to be replaced by platform-specific high precision sleep code later on.