TransportPosition: convert m_sLabel -> m_type
we do not use a lot of position and their label is fixed. As it turns out we can not do a plain comparison of `std::shared_ptr` to determine which of the positions we are dealing with but have to base this check on label/type. And since this one will happen in the `AudioEngine` we do want to compare an enum class and not a QString (efficiency)