MidiActionTest: use fresh song in bar change tests
within our Windows pipeline `testPreviousBar` does routinely fail and it seems to do so because the grid of the song is in a somewhat different state than on other platforms. Maybe execution order of individual unit tests is different there.

Anyway, with this patch the toggling of grid cells is done on a fresh and empty song. This should work on all platforms