split off TimeHelper for high resolution sleep
turns out to do a lot more magic in order to increase accuracy of the high precision sleep variant. This will involve a state, like a vector of measurements in order to calculate statistics, a mutex, a thread to do a burn in and so on. Best to encapsulate all this in a class instead of static members.

The Clock aliases an stringification has been kept in place because it is more light-weighted and used in Object.h