PlayerControl: discard EVENT_TEMPO_CHANGED
in case the current tempo source is either Timeline or an external application.

`PlayerControl::tempoChangedEvent` creates a popup warning about the tempo change. This dialog blocks the handling of all other events until it is closed.

Dialing a button associated with e.g. a BPM_INCR MIDI action results in a lot of `EVENT_TEMPO_CHANGED` events. Previously, each one of them would result in a separate dialog - spamming the user to such a degree Hydrogen would almost become unusable. Now, all remaining events of that type are discarded once the dialog is closed