make Theme std::shared_ptr
resource handling did not went as expected and in result we had a large number of `Theme` (and the individual subsets) created during runtime. At one point I also had a bug were an ever increasing amount of themes were created - a huge memory leak. Unfortunately, I was not able to reproduce this one.

While at it I introduced dedicated getter functions in `Preferences` to retrieve individual parts of the theme and make the `loadFrom` part of the `InterfaceTheme` fall back to the values defined in the constructor