Sampler: harden against invalid notes
when dealing with a massiv number of notes - resolution off and several rows completely filled while drawing with resolution off at the same till - a `Note` in `Sampler::m_playingNotesQueue` can be nullptr. It is an edge case and quite tedious to reproduce. So, I leave it at checking for nullptr and poping the corresponding note from the queue