SampleEditor: use preview instrument on Play button
Whenever the instrument of the sample is muted or another one is soloed, the "Play" button will not produce any sounds anymore. While "Play Original" still works. That was most probably not intended designing the dialog.

Now, the instrument used on the "Play" button is in "preview" state and independent of the kit