rework missing sample handling
Previously, `InstrumentLayer`s with missing samples have been partially discarded while saving. Partially, since they were not written to the resulting song file, the missing samples property of the instrument was cleared, but the layer itself was not touched and could be still accessed.

Instead, the user will now be prompt on saving a song whether to discard or keep all layers with missing samples. This allows for both saving the song without any loss of information and to make the current state of Hydrogen correspond to the saved .h2song file. If opting for discard, the `InstrumentLayer` is actually removed from the current drumkit