SoundLibraryPanel: update widget on song save
to have reflect a possible new song in the user song folder. This could be made more efficiently by just refreshing in case the user song folder is part of the song file path. But it is not done that frequently and I think it is a better solution compared to increasing the complexity by introducing another event type