Instrument: do not mute instruments with missing samples
way in the past - at the beginning of Hydrogen - each instrument typically just held a single sample. Back then it was a convenient UX tweak to mute an instrument in case the sample could not be loaded. But nowadays instruments have multiple layers and can consist of several components holding multiple layers. Muting the whole instrument because a single sample is missing is over the top