Duplicate instruments via popup menu (#1925)
we now support to duplicate instruments and all their layers via the right-click context menu of the sidebar of the pattern editor. The action can be undone.

The notes assigned to the instrument will not be duplicated. This is already possible via another context action. It is probably best to have a larger set of more granular and disjunct actions