PatternEditorSidebar: inline editing of instrument names
Instead of a separate popup dialog querrying the user for the new instrument name, we now support inline editing in the instrument label widget itself. UX is like the one in the pattern list: <kbd>Enter</kbd> and clicking outside make Hydrogen adopt the new name while <kbd>Esc</kdb> discards it.

As a side effect editing the instruments name via the top label of the `InstrumentEditor` is temporarily disabled