Pattern: incorporate legacy load into load_from
and drop `Legacy::loadPattern` instead.

When inspecting a very old song file provided by a user it turned out our previous assumption - old and new patterns can be distinguished by checking whether the name is incorporated via <name> or <pattern_name> - does not hold. Instead, we have to check whether the notes are directly inserted into a <noteList> or wrapped in a <sequenceList> (in versions prior to 0.9.4)