SoundLibraryPanel: fix pattern load
previously, patterns have been loaded by storing two values and compining them later on: the name of the pattern and the name of the drumkit the pattern was created with.

This fragile approach broke when I changed the tool tip of the pattern elements in the SL tree to be more descriptive (since the tooltip was used _as is_ as drumkit name). In addition, this approach does not allow loading arbitrary pattern - one of the main things of version 2.0.

So, in order to fix broken loading and make the 1.2.X release line capable of handling patterns created in version >= 2.0, loading and deleting patterns is now based on their actual path