Shortcuts: unique playlist default shortcuts
the current default shortcuts for the playlist did overlap with the ones of the main application. Before centralizing shortcuts and allowing to customize it, this was ok. But now, e.g. <kbd>Ctrl</kbd> + <kbd>S</kbd> pressed in the `Mixer` would trigger _both_ saving of the current song and the current playlist. This is not intended.

Since these are just default shortcuts the user can alter, I do not think we cause any harm when making them more complex