Preferences: add MidiInput/OutputMapping enums
how incoming notes are mapped to instruments and which MIDI notes the latter do trigger on playback is currently stored via various different settings and it is not easy for the user/developer to determine the current state.

Instead, these two enum classes will encompass all possible perculations as well as introducing stubs for new options which will be implemented soon.