Event: add unique id
a sufficiently unique ID is now assigned to each created `Event` within the `EventQueue`. It can be retrieved as return value of `EventQueue::pushEvent()`.

This ID will allow the GUI to track events is triggered itself. This way it can deliberately decide not to act on than - e.g. when interacting with a widget which triggers an event that would result in the update of the very widget.