MidiInstrumentMap: implement instrument <-> MIDI mapping
for the input mapping there is no single source of truth but two different ways - via a pointer to an `Instrument` and the received MIDI note - to access the mapping. This is unfortunate but largely due to the fact that we so many different ways to map the input.

The fallback input channel for each instrument is the one used for the output.

The fallback input note and channel in custom mapping in case no explicit custom mapping is provided yet are the ones used for the output.

input mapping itself is not properly wired within the core yet