MidiControlDialog: rework action MIDI channel handling
instead of using a combo box, the value for allowed channels for MIDI actions and Clock events is now represented with the same kind of spin box as used for all other MIDI channel widget. This, hopefully, makes configuration more straight forward and intuitive.

The underlying member variable in `Preferences` was renamed `m_nMidiChannelFilter` -> `m_nMidiActionChannel` to more approriately describe what it is used for.

In an effort to align the range of possible values and settings, the input channel for actions and clock events can now be set to "Off" too. That is a little tricky, since it used `-1` to encode "All" channels previously, while the per-instrument MIDI output channels and custom MIDI input channels use `-1` to encode "Off". But it is fixed in a backward compatible way. Though, note that when setting the action channel to "Off" (`-2`), the corresponding value in the `hydrogen.conf` file can only by handled properly by versions of Hydrogen >=1.2.7