implement MIDI output note mapping
this also fixes the issue that note off messages did previously only worked properly when using the default pitch. Any customizations in key or octave resluted in a NOTE_OFF with a wrong note value.

The NOTE_OFF values send for all instruments when stopping playback do still miss but I will keep it as is for now. In order to fix this issue, we would require to keep record of all notes for which we did send a `NOTE_ON` but no `NOTE_OFF` yet. This is feasible but requires to much resources for a feature no one did actually ask for