UndoActions: drop SE_loadPatternAction
in favor for `SE_insertPatternAction`. Previously, a snapshot of the
pattern group vector was saved as a XML file in order to undo the
action. But replacing a pattern with another one should certainly leave
the grid cells of the `SongEditor` untouched. Otherwise, the user would
add a new pattern and/or remove the old.

Inspired by #1269.