PatternEditorSidebar: fix instrument id rendering
This did work at some point. But for whatever reason the margin/text
width we use to derive the position of the vertical separator in the
type label while displaying the instrument ID was off.

This was now fixed by introducing a magic number as an additional
offset. Not pretty at all. But I do not see what else to do. It seems to
be independent of the Qt version and font to use