CoreActionController: fix event in setPattern
instead of `Event::Type::PatternModified` we do now queue
`Event::Type::SelectedPatternChanged` because both adding a pattern and
replacing an existing one will cause the currently selected pattern to
be changed. Without the proper event the DB of the `PatternEditorPanel`
won't be updated and unmapped notes won't be displayed properly.