InstrumentLayer: expose setSample() in Instrument
and make the version of the layer class itself private.

It was an antipattern that parts of the GUI went deep done into a
drumkit and changed individual samples or layers.

Now `Instrument::setSample()` is the central entry point for changing
samples. This allows for checking for missing ones on the fly.