Instrument: trigger event on missing sample change
Now, whenever a missing sample of an instrument is replaced by a new
one, the missing sample icon in the sidebar of the `PatternEditor` will
disappear.

Note that the actual event used is not the final one. This will be dealt
with in an upcoming commit as it most probably requires a new event.