Rename InstrumentRack -> Rack/Rack
Since the rack does not hold only instrument-specific things - as the
sound library -, `Rack` is a more appropriate name. Especially, since it
does now also include the ComponentsEditor.

It has been moved in a dedicated folder which will contain (in future
commits) all widgets which are children of the rack.