InstrumentComponent: drop max layers
We already started porting the instrument layers to a vector of
arbitrary size instead of an array of predefined length filled with
`nullptr` for "missing" layers. In this commit, the whole notion of
having a maximum number of layers is dropped.

Instead, the user can now add an arbitrary number of layers to a
component (GUI will be adopted in upcoming commits).