Split of LayerWaveDisplay from WaveDisplay
Instead of having a widget-like class `WaveDisplay` in
`Rack/ComponentEditor` and inheriting it in other parts of the
application, it is now moved into `Widgets/`.

In `Rack/ComponentEditor` a minimal wrapper called `LayerWaveDisplay`
was introduced instead.

This will allow us to further customize the layer wave display in the
`ComponentView` - e.g. adding drag & drop capabilities - without having
to worry about breaking things in `PlaybackTrackWaveDisplay`.