Fix initial instrument type assignment
Previously, when assigning the initial instrument type to an instrument
`DrumkitPropertiesDialog` created an undo macro for

1. changing the types of all notes associated with the instrument.
2. changing the instrument type using
`UndoActions::SE_setInstrumentTypeAction`.
3. replacing the current drumkit with a new one.

But this caused issues when undoing the instrument type assignment.

The reason behind the inconsistencies was step 2. calling
`CoreActionController::setInstrumentType`, which acted on the particular
instrument of the song's _current_ drumkit. Not the one we intended to
switch to in step 3.

A bunch of instrument type setting code has been dropped and the types
are now assigned directly to the new drumkit in `DrumkitPropertiesDialog`.