add SoloButton and MuteButton
add dedicated widgets for solo and mute button.

While other buttons are (still) not flat, these ones already are (looks
more modern and better integrate the default buttons in `QToolBar`).

Both buttons have a non-standard colorful background derived from their
color used when checked. This is done in order to make them both
distinct and more quickly to discover.

In general, this could be implemented with another `Button::Type`. But I
choose not to because these serve very specific purposes and it is very
unlikely we want to reuse them in another widget.