SampleTest: fix drumkit handling
This one did occasionally break the macOS and Windows pipeline. Turns
out cppunit - the framework we use for unit testing - does store all
registered test suites in a `std::set` and they are not guaranteed to
be run in the order of registration. On my local devices they are. But
within the pipeline obviously not...

During the test the original kits within the sound library were
modified. When run _prior_ to `NoteTest` note mapping tests would fail
because the assumptions within the test would not hold anymore.

Now, deep copies of the kits are used instead.