MidiEventMap: use std::vector as basis
Previously, MIDI Note, CC, PC, and MMC have been stored in different
maps/classes. But this had a conceptional flaw: both the format in the
config file as well as the rendering of the events in the GUI featured
an order. In `MidiEventMap`, on the other hand, such order could not be
applied. Instead, when listing the individual events in
`MidiActionTable` they were grouped by type - as those individual data
structures were querried sequentially.

Thus, the order of events assigned by the user might get mixed up.

Instead, we now use a plain `std::vector` to store _all_ events.
Searching for events when handling incoming MIDI messages might be a
little slower. But unless the user has a massive amount of mappings -
which is not expected - this shouldn't be noticable.