ColoredButton: use default color over gradient
The radial gradient did not felt right. It did not worked with the flat
UI of recent updates at all.

Instead, it was dropped and a default background color is set for
unchecked buttons instead. The corresponding text color is the highlight
color `m_baseColor` of the button. Since can (and does) give poor
contrast, we used a bold font weight for the button.

I also experimented with a black outline of the text. But the result was
poor and it would come with costs, like losing the ability to easily
center the rendered text.