NotePropertiesRuler: override getElementsAtPoint
while note interaction worked fine for most of the properties in
`NotePropertiesRuler`, doing so in the key/octave view felt really
awkward. Being unable to select individual notes while they are rendered
separately in dedicated rows/columns does not integrate well with the UX
of `DrumPatternEditor` and `PianoRollEditor`.

Instead, `NotePropertiesRuler` does now reimplement
`getElementsAtPoint()` in order to allow to select them individually in
the key/octave view.

One thing that still yields inconsistent UX is lasso selection. In
key/octave the user probably expect now to be able to lasso-select
individual notes in one column. However, implementing this would be a
real pain. And since it is quite an edge case, I choose to leave it as
is for now.