PatternEditor: enforce action in handleElements
The function does now check whether it should add or delete something.

In addition, adding new notes is now only possible if there is no other
note at the corresponding grid point.

Prior to this commit a note could be deleted by either
hovering and clicking it or by clicking its grid cell. The latter option
was kept to ensure a certain degree of backward compatibility with the
UX found in version 1.2.X. There the user could only toggle notes at
grid points and not directly interact with them.

But the above behavior proofed problematic once we allowed to
create/delete notes in the `NotePropertiesRuler`. There it would be just
too easy to accidentally delete notes. And disabling grid-based deletion
just in there would result in a more inconsistent UX. So, it has to go
away altogether.